+++ app/src/main/java/utn/frba/mobile/moodmatch/MainActivity.kt
@@ -27,7 +27,7 @@ class MainActivity : ComponentActivity() {
 @Preview(showBackground = true)
 @Composable
 fun MoodMatchApp(modifier: Modifier = Modifier) {
-//    InitialScreen("Android")
+    //InitialScreen("Android")
     //SignInScreen()
     RecommendationScreen()
 }
